verneuil/replication_buffer: use regular NamedTempFile for chunks
We used to rely on O_TMPFILE + linkat via `/proc/self/fd` to publish
new chunk files without leaking anything on crashes or program
shutdown.  That idiom is hard to emulate outside on !Linux platforms,
and we already have logic to clean up stale temporary files.

Use a `NamedTempFile` and a standard path-based hard link: that works
just as well and is more portable.

Part of Txxxx.