verneuil/verneuilctl: parse hostname and path from a verneuil:// "uri"
Summary:
We'll want to offer that flexibility for the read replica.

The URI scheme is `verneuil://hostname/path/to/db`, or
`verneuil:///path/to/db` for a db replicated from the current host.

In preparation for Txxxx.

Test Plan: verneuilctl restore on `verneuil://machine.hostname.com/path/to/sqlite.db`

Reviewers: ahicks, jshufro

Reviewed By: jshufro

Differential Revision: https://foo.backtrace.io/Dxxxx