verneuil/snapshot_vfs_ops: reuse pre-existing snapshots if possible
Summary:
We expect applications to open multiple connections for the same
manifest URI.  It makes sense to deduplicate the snapshots as much as
possible, and to make updates explicit.

Start by deduplicating snapshots, and always reusing the current
snapshot for a manifest URI, if there is one.

Depends on Dxxxx.

Part of Txxxx.

Test Plan: manual in sqlite3 shell, with `eprintln!` in `fetch_new_data` / `get_data`.

Reviewers: jshufro

Reviewed By: jshufro

Differential Revision: https://foo.backtrace.io/Dxxxx