verneuil/snapshot_vfs_ops: track lock level for replica snapshots
Summary:
We don't want to change the snapshot data under sqlite3's feet when it
holds a read lock.  Track whether sqlite3 currently believes it's
holding a read lock on the snapshot in a bool field: when the read lock
is held, we will never want to update the snapshot data pointer.

Depends on Dxxxx.

In preparation for reload pragmas in Txxxx.

Test Plan: Manual tests in the sqlite3 shell with eprintln!.

Reviewers: jshufro

Reviewed By: jshufro

Differential Revision: https://foo.backtrace.io/Dxxxx