verneuil/snapshot_vfs_ops: new non-blocking reload pragma
Summary:
Executing `pragma verneuil_snapshot_async_reload;` will either refresh
the connection's snapshot data or, if the connection is already up to
date, ensure a background force reload task is in flight.  The pragma
returns `true` if a fresher snapshot was available, and `false`
otherwise, if a background reload task had to be scheduled.

Does not refresh the snapshot if the connection is currently in a read
transaction.

Closes Txxxx.

Test Plan: manual at the sqlite3 shell.

Reviewers: jshufro

Reviewed By: jshufro

Maniphest Tasks: Txxxx

Differential Revision: https://foo.backtrace.io/Dxxxx