verneuil/copier: remove the logic to GET before potential overwrites
We used to issue a GET before any PUT: rumour has it some old versions
of S3 had better back-to-back PUT latency with that pattern.

Many changes have happened since that old version (strong
read-after-write consistency in particular), and stats for the past
week shows the same number of slow GETs as slow PUTs, with similar
latency distributions.  It seems likely that the rate of slow API
calls is identical for our GETs and PUTs, and that issuing GETs
before potential overwriting PUTs merely increases the probability
of a slow API call.

Stop doing that; we can reinstate if stats show a lot more slow PUT
calls after this commit.

Ref Txxxx.