verneuil/snapshot_vfs_ops: track when we fetched a snapshot
Summary:
Each snapshot VFS file refers to shared snapshot data.  It can be
interesting to know when that snapshot was constructed.  For example,
we might want to know how long it's been since we fetched the snapshot
data's manifest: if the manifest is several minutes old, we might want
to see if a new one is available.

Follow-up commits will expose that information.

Part of Txxxx.

Test Plan: Follow-up commits.

Reviewers: jshufro

Reviewed By: jshufro

Differential Revision: https://foo.backtrace.io/Dxxxx