verneuil/snapshot_vfs_ops: make force refresh interact with async reload
Summary:
The `verneuil_snapshot_refresh` pragma now accepts a ternary argument:

- no argument, 0, or falsey: update to the latest available snapshot
  without blocking

- 1, or truthy: update to the latest available snapshot, then fetch
  and update to the result, if no fetch is currently in progress for
  the latest snapshot.

- 2 or "force": always fetch a new snapsnot.

We don't want to block force refresh if an async reload is in
progress, but we can always signal that a fetch is in progress.

Makes Txxxx schedule with async reloads more nicely.

Test Plan: manual testing in sqlite3

Reviewers: jshufro

Reviewed By: jshufro

Differential Revision: https://foo.backtrace.io/Dxxxx