NFC loader: hide Chunk::payload behind a getter
Summary:
We'll want to store an `memmap2::Mmap`, instead of a `Box<[u8]>`.
Let's hide that implementation detail with a getter that returns a
`&[u8]`.

Also drop the `Eq` / `PartialEq` traits: `Mmap` don't implement
`PartialEq`, and we don't compare Chunks for equality.

In preparation for using file-backed mappings in Txxxx.

Test Plan: sqlite tests.

Reviewers: jshufro

Reviewed By: jshufro

Differential Revision: https://foo.backtrace.io/Dxxxx