verneuil/manifest_schema: be less noisy about new db files.
When sqlite creates a new db file, we expect not to find any version
id: we set that xattr on demand.  We also don't really care about
blips in missing xattrs, they only create problems when they're
consistently missing.

Keep the `warn`ing actionable and only trigger that when we fail to
read the xattr on an old file: that's more likely to be a consistent
problem.

Ref Txxxx.