verneuil/copier: lower log level for ctime mismatch when headers match
While it is true that we've seen sqlite change the data on disk
without changing the header's version counter, we do have logic to
catch such cases.  Given this additional protection (exercised in
sqlite tests), it seems likely that, when we notice a sqlite file's
ctime differs from the latest replication data, but the sqlite
headers match, we're looking at a false positive.

For example, we recently observed a burst of such "mismatches" for ~5%
of sqlite dbs where we had triggered empty immediate transactions, and
confirmed that the sha256 of each local db disk did match its
replica's.

Still log about this condition, but only at debug level.

Ref Txxxx.