verneuil/manifest_schema: do not mix the ctime in the version id
at least not by default.

It looks like no-op write transactions can sometimes result in updates
to ctime without any change to the file data.  When that happens, we
don't want to force re-replicate the whole db file.

Closer examination of the sqlite source code confirmed that such
silent changes happen when the changed data is only on
unused (recyclable) pages, or when sqlite downgrades its write lock
after failing to fully roll back a failed transaction (a simulated
failure scenario in tests).

In the former case, it's not necessary to replicate the new state of
the unused page: it will be written before it becomes used, and we
will then replicate its contents.  In the latter case, the db file
might actually be in an inconsistent state, and writes will be rolled
back at the beginning of the next transactions.  We don't want to
replicate failed write transactions, so the temporary mismatch is OK.

Ref Txxxx.