verneuil/tracker: do not snapshot db files that have hot journals
If there's a hot journal when downgrading the lock to SHARED or lower,
all writes to the db are about to be rolled back, so we don't want to
track them: not only are the changes transient, they might also
represent an invalid database!

Sqlite attempts to roll back before releasing the lock, so this
situation should only happen after catastrophic I/O failure.

Ref Txxxx.