replication_buffer: irreversibly shorten long paths
When we mangle paths to make them fit in file names, we now check
if the result is close to our estimate of the filesystem's filename
limit.  If so, we construct a shorter filename by replacing the
middle of the long mangled name with a sha256 hex digest of the
original mangled path; the result is always comfortably shorter
than 255 bytes.

We can now look at the length of each mangled path (as a filename),
and determine if we're supposed to be able to reverse the mangling:
if the mangled string is at most MAX_MANGLED_NAME_LENGTH bytes,
it's reversible, otherwise it isn't.  We also avoid ambiguity when
encoding by truncating long mangled strings to a bit more than
MAX_MANGLED_NAME_LENGTH, and by making sure reversible strings are
always shorter than MAX_MANGLED_NAME_LENGTH.

Fixes #1.

TESTED=new unit tests.