NFC manifest_schema: apply an optional base byte array to chunks array
The array of chunk fingerprints in manifest protos can be long and is
rarely compressible.  However, we also know they don't change a lot,
so we should be able to expose redundancy for compressors.

Simply apply a xor filter to that array, with a base sequence of bytes
stored in a regular content-addressed chunk, when such a base chunk
is defined in the manifest proto.

This is currently a no-op because we never populate the field and
we also never pass a list of `ReplicationTarget`s to fetch the
base chunk from.

Implements phase 1 of #3.

TESTED=sqlite tests.