replication_buffer: read manifests with a list of replication targets
That functionality is only used by the change `tracker.rs`, and it
always has such a list on hand.

Moreover, when reading the current staged manifest, make sure to try
and load from the staged chunk directory: we should find any chunk
needed to decode the manifest in there.

We also have to relax the invariant checking logic when the validator
might be racing with a snapshotter: when that happens, we could read a
manifest file, and not find the corresponding chunks.  We can only
rely on the staged manifest being readable when we hold a lock (read
or write) on the sqlite db.

Makes phase 1 of #3 correctly configurable.

TESTED=follow-up commits.