loader: increase chunk size limit
When we also upload "base" arrays of fingerprints as chunks, we may
send larger chunks than 64 KB.  Verneuil already manages 3-4 GB
databases, which translates into base chunks of up to 512 KB, and
that's uncomfortably close to the 1 MB limit on chunk size.

Bump that up to a bit more than 256 MB, which would be enough even
for a 1 TB database (although we'd likely want a different approach
to metadata at that size).

In preparation for #3.