NFC unzstd: new module to sniff and decompressed zstd-compressed bytes
If only because of backward compatibility, we have to try and detect
when data is definitely not compressed, and avoid reporting that as a
decoding error.  Callers can then decide if they want to decode the
raw or decompressed data.  In practice, our data is unlikely to
accidentally include a zstd header, so we can assume we want to
propagate decoding errors.

In preparation for zstd-compressed manifest blobs (step 3 of #3).

TESTED=sqlite tests.