manifest_schema: see if bytes must be unzstd-ed before proto parsing
If the bytes don't begin with the zstd header, directly parse them as
proto.

If the bytes begin with the zstd header and can be decompressed, only
attempt to parse the decompressed bytes: it seems highly improbable
that a valid proto message would also be a valid zstd stream,
including checksums.

Otherwise, the bytes begin with a zstd header but fail decompression.
Give decoding the raw bytes (as proto) a shot; if that fails, report
the decompression failure.

Making sure field 5 in `Manifest` isn't used as a varint guarantees
our proto messages don't start with the zstd header, which avoids the
last ambiguous case, where we try to zstd decompress, fail, and try to
parse the raw bytes as a proto.  Document that restriction on future
schemas.

Implements step 3 of #3.

TESTED=sqlite tests.