copier: minimally compress manifests on upload
We can now expect some compressibility when we use a base chunk for
the list of chunk fingerprints.  Try to take advantage of that with
quick compression.

We now unconditionally ask zstd to compress manifest blobs as quickly
as possible: in the worst case, we take a 12 byte hit for zstd
framing.  That seems negligible, and protobuf bytes tend to be a
little compressible, so the actual overhead is less than that.

Also rename the compression level constants to be clearer about
intent.

Implements step 4 of #3.

TESTED=sqlite tests, looked for compressed manifests in minio during tests.