copier: compress large blobs quickly
When a content-address chunk is larger than the snapshot
granularity (chunk size), it's probably not snapshot data..
which means it's probably a base list of fingerprints.

Either way, it's large, and we probably don't want to waste too much
time compressing it.  When we're asked to compress at the default
level and find such a large chunk, instead compress at the minimum
quality (maximum speed).

Improves the upload speed of large base fingerprint chunks (for #3).

TESTED=sqlite tests?