tracker: implement a baseline base chunk reuse policy
We essentially want to keep using a base fingerprint chunk while it's
still a good match for our actual list of chunk fingerprints.  Simply
reuse with probability proportional to the match rate.

On top of that, when we don't want to reuse the current chunk, look at
the size of the list of fingerprints: it's not useful to indirect
through a base chunk when the list is short anyway.  For now, set an
arbitrary cutoff at 1024: if a manifest serialises to ~16 KB or less,
don't try to make it compress better.

We only take the list size into account when we're about to create a
new base chunk, instead of attempting to add some hysteresis: we're
not worried about rapid alternation between having and not having a
base chunk when the database goes between 1023 and 1024 chunks because
we only "lose" a base chunk stochastically, when it's not useful.

Closes #3, by implementing the policy (steps 6 and 7).

The test-only random policy shows:
1. manifests with no base chunks
2. manifests with a fresh base chunk
3. manifests with a fully stale base chunk
4. manifests with a partially stale base chunk

TESTED=sqlite tests for lack of crash, manual inspection for impact during tests.