copier: treat fingerprints for bundled chunks specially
`CopierWorker`s are also responsible for keeping relevant chunks
alive: given a manifest, they figure out the list of chunks the
manifest references, and periodically "touch" them (by copying each
blob over itself in S3).

When something goes wrong in that "touch" process, we assume the
manifest is now useless, and the source database's replication state
is cleared to force a full snapshot from scratch.

We don't want that to happen whenever we try to touch a chunk that was
bundled with the manifest instead of uploaded to S3.  Tweak the list
of chunks in the manifest to treat bundle chunks like the well-known
zero fingerprint, and not patrol touch them.

Avoids degrading performance once #7 introduces expected "missing" chunks.

TESTED=follow-up commits?