tracker/invariant: check `consuming` snapshot harder
We already check the `consuming` snapshot, and then the `ready` one.
There is a timing window where we will find nothing in `consuming`,
a `Copier` will move `ready` to `consuming, and we will then also
find nothing in `ready`.

Work harder to validate in-flight snapshots by validaing the
`consuming` snapshot before *and after* validating the `ready`
snapshot.

TESTED=it's a new test assertion.