manifest_schema: extend the V1 proto with bundled chunks
We want to bundle the chunk at offset 0 (that contains sqlite's header
page), because that chunk includes a counter that's incremented after
every write transaction; we don't expect a lot of deduplication from
content addressing.

This commit extends the schema and documents the new field.  Further
commits will prepare readers for bundled chunks and actually bundle
the first chunk with the manifest.

This schema change lets us implement #7.

TESTED=it builds.