NFC replication_buffer: extract filename shortening function
We generate filenames from paths, and the result can be longer than
the maximum filename size.  We have a function that steers clear of
the industry standard 255 bytes (or character) limits by replacing
the middle of long names with a fingerprint of the full name.

Extract it, so we can use it in more place.