manifest_schema: represent `ManifestV1::generated_by` as a byte array
This is wire compatible with strings, the only difference is we don't
assume the bytes represent valid utf-8.  We still generate valid
utf-8, so there is no observable difference in the proto bytes we
generate; however, we now accept messages in which the `generated_by`
field isn't valid utf-8.

While we're at it, drop the `Option` in the Rust struct: the default
empty value is fine.

TESTED=updated unit tests.