replication_buffer: scan for spooling subdirectories on startup
We derive spooling subdirectory names from the current boot time and
boot id.  The boot time is subject to time adjustments, so we will
sometimes generate a different spooling subdirectory than an older
process.  That's not bad for correctness, but can result in useless
replication work.

The boot id ensures correctness of our name generation scheme; the
boot time is only part of the name to help administrators.  Scan for
subdirectories with similar boot times to the one we just found when
constructing spooling subdirectories.

TESTED=sqlite tests?