snapshot: prepare for lazily-loaded chunks
In order to load chunks on demand, we have to store enough information
to load the chunks after the Snapshot has been created, and also try
to detect when that must be done.

Stick that information in `LazyChunk`s, and make sure they're loaded
before reading their contents...  but always populate `LazyChunk`s on
creation for now.

Part of #12.

TESTED=existing tests.