NFC snapshot: prepare for not loading all chunks in `Snapshot::new`
We now check that we did fetch all the chunks we tried to fetch in one
pass, and populate the snapshot with what we can find in a second
pass.

We currently always fetch everything, so there's no actual difference
in behaviour, except maybe the order in which we check that everything
looks good.

Part of #12.

TESTED=sqlite tests.