snapshot: implement the partial loading policy
We want to load the first `~sqrt(n)` chunks for a db that consists of
`n` chunks.  That should hit most of the internal b-tree pages, and
scales nicely with the size of the database.

Of course, when the db is small, lazy loading isn't super useful, so
we also clamp that from below.

In all cases, we must fetch the last page eagerly: it could be shorter
than the `SNAPSHOT_GRANULARITY`, and we assume that all
yet-to-be-loaded chunks are exactly that size.

Implements the internal bits of #12.  Now to make it configurable.

TESTED=sqlite tests.