HACK tracker: avoid nested runtimes by spawning short-lived threads
`copy_spool_path` is a sync function that uses async code internally,
with a thread-local `current_thread` runtime.  That works great when
called from C or sync threads, but tokio really doesn't like nested
runtimes.  Detect when the caller is already in a tokio async context,
and spawn a short-lived thread to handle the copying.

That's a bit unfortunate, but, realistically, the overhead of spawning
and tearing down a thread is probably marginal compared to uploading
blobs to remote storage.

Closes #15.

TESTED=manually.