loader: run `load_from_source` in the Rayon `LOADER_POOL`
The implementation (`load_from_source_impl`) uses the
Verneuil-internal current-thread tokio runtime, and that will panic
when the caller is already in an async context.  We already try to run
all data loading logic in the `LOADER_POOL`; we might as well
guarantee that `load_from_source_impl` *always* runs on that pool.

The one small hitch is that `load_from_source` is already called from
that thread pool.  We just have to detect when we're running on the
pool and not try to re-enter it.

Simplifies transitive callers in Verneuil's public interface for #16.

TESTED=sqlite tests.