c/vfs: add support for Darwin's 32-bit dev/ino
The fields are opaque to the Rust side; the only thing that matters
is that they each take up one u64's worth of space in the LinuxFile
(sic) struct. Achieve that by union-ing the field with uint64_t,
and relax the static assertion to sizeof(dev_t) <= sizeof(uint64_t),
etc.