process_id: improve fallback path when process birth time is missing
We want to disambiguate pids to avoid collisions when pids are
recycled. Some OSes have explicit support for unique pids, but
that's not super widespread, so we fake it by augmenting the pid
with the process's birth timestamp (with arbitrary units and epoch).

Unfortunately, it seems that Darwin now only exposes btime
(or its native unique process identifier) to privileged processes,
so we have to improve the fallback logic.

The old implementation simply substituted the constant `u64::MAX`
for the btime when it was missing.  The new implementation
generates a random u32.

Of course, this means the "btime" associated with a given process
may not vary.  That's why this commit introduces AtomicKV32, an
atomic register for key-value pairs of u32, where Last Write Wins
when the key changes, but First Write Wins when the key doesn't.

The process_id module uses this AtomicKV32 to associate a consistent
btime to a pid, while still updating the `(pid, btime-or-random-u32)`
pair when the pid changes (after fork).

The value must be a u32, so we now also truncate the btime, when
we have one, to 32 bits.  That's enough to disambiguate pids.

While we're working on !Linux platforms, only try to get the btime
for pid 1 on Linux.