manifest_schema: unit test xattr versioning machinery
Verneuil tags SQLite DB files with a uuid stored in the
`user.verneuil.version_id` xattr.  This makes it possible to detect
writes when SQLite doesn't update the version field in the DB file's
header (typically when rolling back a transaction that wrote to
pages that are logically unused).

That's probably not important in practice, but helps us test replication
by comparing raw bytes.  Still, test the xattr machinery before
adding support for the BSD-style calls.