process_id: handle malicious comm fields
I don't think there's an attack vector here, but we might as well
get this right. Like https://www.openwall.com/lists/oss-security/2022/12/21/6
points out, the comm field is a freeform string; the saving grace
for `/proc/pid/stat` is that it's the only such string, and it's
the second field.  We can thus disregard everything until the
first parenthesis... and if we find multiple matches over multiple
lines, we want the last matching line.

And add a new test case for space in the comm field.