tracker/snapshot_file_content: avoid useless uploads after desync
When the replicating (writer) Verneuil VFS notices that there is
spooled metadata, but that it doesn't match the sqlite file, we
currently always copy everything from scratch.  This makes sense
for long-lived processes, but isn't great (still correct, just
wasteful!) for short-lived ones that run between regular writes
to the sqlite db.

This commit adds logic to notice when we have a valid spooled manifest,
but it doesn't match the file on disk.  We then force a full scan of
the sqlite file, but still avoid uploading any chunk that happens
to match what was in the manifest.

This new feature does not assume any relationship between the file
on disk and the manifest we happen to find in the spooling directory.
The only assumption is that chunks in that manifest have been
uploaded, or are staged for upload.  If any chunk in the manifest
happens to match the corresponding page in the sqlite file, that's
great, if not, we just upload the new data, with content addressing.

The old behaviour (upload everything) is preserved when we don't
find a manifest, e.g., because the spooling directory is empty, or
because the manifest predates the most recent boot.

Independently, it might be interesting to treat the previous manifest
as a set of chunks that are known to be already uploaded or staged
for upload, and not stage them again.  In practice, we mostly expect
matches at identical page offset, but who knows with large blobs?